PHPStan Changes

+++ src/Entity/Offers.php
@@ -364,7 +364,7 @@ class Offers
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
